… aten/src/ATen/native/vulkan/ops/Concat.cpp
… aten/src/ATen/native/vulkan/ops/Convolution.cpp
… aten/src/ATen/native/vulkan/ops/Convolution.h
… aten/src/ATen/native/vulkan/ops/Copy.cpp
… aten/src/ATen/native/vulkan/ops/Gru.cpp
… aten/src/ATen/native/vulkan/ops/Mean.cpp
… aten/src/ATen/native/vulkan/ops/Mm.cpp
… aten/src/ATen/native/vulkan/ops/Mm.h
… aten/src/ATen/native/vulkan/ops/Padding.cpp
… aten/src/ATen/native/vulkan/ops/Permute.cpp
… aten/src/ATen/native/vulkan/ops/Pool.cpp
… aten/src/ATen/native/vulkan/ops/Shape.cpp
… aten/src/ATen/native/vulkan/ops/Slice.cpp
… aten/src/ATen/native/vulkan/ops/Softmax.cpp
… aten/src/ATen/native/vulkan/ops/Tensor.cpp
… aten/src/ATen/native/vulkan/ops/TransposeConvolution2d.cpp
… aten/src/ATen/native/vulkan/ops/Upsample.cpp
… aten/src/ATen/test/vulkan_api_test.cpp
… aten/src/ATen/test/vulkan_perf_test.cpp